+++ src/ska_sdp_spectral_line_imaging/__init__.py
@@ -2,6 +2,6 @@
 
 """Spectral line imaging pipeline package"""
 
-__version__ = "0.6.1"
+__version__ = "0.6.2"
 
 __author__ = "Team DHRUVA"